# backend/YanCarz/swagger.json
{
    "openapi":  "3.0.4",
    "info":  {
                 "title":  "YanCarz.API",
                 "version":  "1.0"
             },
    "paths":  {
...
                  "/api/shared/Mark":  {
                                           "get":  {
                                                       "tags":  [
                                                                    "Mark"
                                                                ],
                                                       "responses":  {
                                                                         "200":  {
                                                                                     "description":  "OK",
                                                                                     "content":  {
                                                                                                     "text/plain":  {
                                                                                                                        "schema":  {
                                                                                                                                       "type":  "array",
                                                                                                                                       "items":  {
                                                                                                                                                     "$ref":  "#/components/schemas/MarkDto"
                                                                                                                                                 }
                                                                                                                                   }
                                                                                                                    },
                                                                                                     "application/json":  {
                                                                                                                              "schema":  {
                                                                                                                                             "type":  "array",
                                                                                                                                             "items":  {
                                                                                                                                                           "$ref":  "#/components/schemas/MarkDto"
                                                                                                                                                       }
                                                                                                                                         }
                                                                                                                          },
                                                                                                     "text/json":  {
                                                                                                                       "schema":  {
                                                                                                                                      "type":  "array",
                                                                                                                                      "items":  {
                                                                                                                                                    "$ref":  "#/components/schemas/MarkDto"
                                                                                                                                                }
                                                                                                                                  }
                                                                                                                   }
                                                                                                 }
                                                                                 }
                                                                     }
                                                   },
                                           "post":  {
                                                        "tags":  [
                                                                     "Mark"
                                                                 ],
                                                        "requestBody":  {
                                                                            "content":  {
                                                                                            "application/json":  {
                                                                                                                     "schema":  {
                                                                                                                                    "type":  "string"
                                                                                                                                }
                                                                                                                 },
                                                                                            "text/json":  {
                                                                                                              "schema":  {
                                                                                                                             "type":  "string"
                                                                                                                         }
                                                                                                          },
                                                                                            "application/*+json":  {
                                                                                                                       "schema":  {
                                                                                                                                      "type":  "string"
                                                                                                                                  }
                                                                                                                   }
                                                                                        }
                                                                        },
                                                        "responses":  {
                                                                          "200":  {
                                                                                      "description":  "OK"
                                                                                  }
                                                                      }
                                                    }
                                       },
                  "/api/shared/Mark/{id}":  {
                                                "get":  {
                                                            "tags":  [
                                                                         "Mark"
                                                                     ],
                                                            "parameters":  [
                                                                               {
                                                                                   "name":  "id",
                                                                                   "in":  "path",
                                                                                   "required":  true,
                                                                                   "schema":  {
                                                                                                  "type":  "string",
                                                                                                  "format":  "uuid"
                                                                                              }
                                                                               }
                                                                           ],
                                                            "responses":  {
                                                                              "200":  {
                                                                                          "description":  "OK",
                                                                                          "content":  {
                                                                                                          "text/plain":  {
                                                                                                                             "schema":  {
                                                                                                                                            "$ref":  "#/components/schemas/MarkDto"
                                                                                                                                        }
                                                                                                                         },
                                                                                                          "application/json":  {
                                                                                                                                   "schema":  {
                                                                                                                                                  "$ref":  "#/components/schemas/MarkDto"
                                                                                                                                              }
                                                                                                                               },
                                                                                                          "text/json":  {
                                                                                                                            "schema":  {
                                                                                                                                           "$ref":  "#/components/schemas/MarkDto"
                                                                                                                                       }
                                                                                                                        }
                                                                                                      }
                                                                                      }
                                                                          }
                                                        },
                                                "put":  {
                                                            "tags":  [
                                                                         "Mark"
                                                                     ],
                                                            "parameters":  [
                                                                               {
                                                                                   "name":  "id",
                                                                                   "in":  "path",
                                                                                   "required":  true,
                                                                                   "schema":  {
                                                                                                  "type":  "string",
                                                                                                  "format":  "uuid"
                                                                                              }
                                                                               }
                                                                           ],
                                                            "requestBody":  {
                                                                                "content":  {
                                                                                                "application/json":  {
                                                                                                                         "schema":  {
                                                                                                                                        "$ref":  "#/components/schemas/MarkDto"
                                                                                                                                    }
                                                                                                                     },
                                                                                                "text/json":  {
                                                                                                                  "schema":  {
                                                                                                                                 "$ref":  "#/components/schemas/MarkDto"
                                                                                                                             }
                                                                                                              },
                                                                                                "application/*+json":  {
                                                                                                                           "schema":  {
                                                                                                                                          "$ref":  "#/components/schemas/MarkDto"
                                                                                                                                      }
                                                                                                                       }
                                                                                            }
                                                                            },
                                                            "responses":  {
                                                                              "200":  {
                                                                                          "description":  "OK"
                                                                                      }
                                                                          }
                                                        },
                                                "delete":  {
                                                               "tags":  [
                                                                            "Mark"
                                                                        ],
                                                               "parameters":  [
                                                                                  {
                                                                                      "name":  "id",
                                                                                      "in":  "path",
                                                                                      "required":  true,
                                                                                      "schema":  {
                                                                                                     "type":  "string",
                                                                                                     "format":  "uuid"
                                                                                                 }
                                                                                  }
                                                                              ],
                                                               "responses":  {
                                                                                 "200":  {
                                                                                             "description":  "OK"
                                                                                         }
                                                                             }
                                                           }
                                            },
...
              },
    "components":  {
                       "schemas":  {
...
                                       "MarkDto":  {
                                                       "type":  "object",
                                                       "properties":  {
                                                                          "id":  {
                                                                                     "type":  "string",
                                                                                     "format":  "uuid"
                                                                                 },
                                                                          "name":  {
                                                                                       "type":  "string",
                                                                                       "nullable":  true
                                                                                   }
                                                                      },
                                                       "additionalProperties":  false
                                                   },
...
                                   }
                   },
...
}